New method "reupload-snapshot-file" that takes the snapshot files stored in local cache and uploads them to the storage
This is needed because the local files in cache are unencrypted, but the
ones in the storage are encrypted. You can't just copy from one location
to the other - the duplicacy encryption routines need to be used.

This code just looks in the local cache folder, finds all revision files
for the specified storage and default snapshot ID and then uploads any
that don't already exist in the remote storage.

This is a problem solving kind of fix for a very specific problem -
where I deleted my storage snapshot files, but still had them cached
locally on my machine.